+++ smoke-tests/apps/LiveMetrics/src/smokeTest/java/com/microsoft/applicationinsights/smoketest/LiveMetricsTest.java
@@ -16,8 +16,20 @@ import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.TOMCA
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.TOMCAT_8_JAVA_8_OPENJ9;
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.WILDFLY_13_JAVA_8;
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.WILDFLY_13_JAVA_8_OPENJ9;
+import static org.assertj.core.api.Assertions.assertThat;
 
+import com.azure.json.JsonProviders;
+import com.azure.json.JsonReader;
+import com.azure.monitor.opentelemetry.autoconfigure.implementation.quickpulse.swagger.models.DocumentIngress;
+import com.azure.monitor.opentelemetry.autoconfigure.implementation.quickpulse.swagger.models.DocumentType;
+import com.azure.monitor.opentelemetry.autoconfigure.implementation.quickpulse.swagger.models.Exception;
+import com.azure.monitor.opentelemetry.autoconfigure.implementation.quickpulse.swagger.models.MetricPoint;
+import com.azure.monitor.opentelemetry.autoconfigure.implementation.quickpulse.swagger.models.MonitoringDataPoint;
+import com.azure.monitor.opentelemetry.autoconfigure.implementation.quickpulse.swagger.models.Trace;
+import java.io.IOException;
 import java.time.Duration;
+import java.util.ArrayList;
+import java.util.List;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -29,12 +41,135 @@ abstract class LiveMetricsTest {
 
   @Test
   @TargetUri("/test")
-  void doMostBasicTest() throws Exception {
-    testing.getTelemetry(0);
-
+  void testTelemetryDataFlow() throws java.lang.Exception {
     Awaitility.await()
-        .atMost(Duration.ofSeconds(10))
-        .until(() -> testing.mockedIngestion.isLiveMetricsPingReceived());
+        .atMost(Duration.ofSeconds(60))
+        .until(() -> testing.mockedIngestion.getCountForType("RequestData") == 1);
+
+    PostBodyVerifier postBodyVerifier = new PostBodyVerifier();
+
+    assertThat(testing.mockedIngestion.isPingReceived()).isTrue();
+
+    List<String> postBodies = testing.mockedIngestion.getPostBodies();
+    assertThat(postBodies).hasSizeGreaterThan(0); // should post at least once
+
+    for (String postBody : postBodies) {
+      postBodyVerifier.searchPostBody(postBody);
+    }
+
+    assertThat(postBodyVerifier.hasExceptionDoc()).isTrue();
+    assertThat(postBodyVerifier.hasTraceDoc()).isTrue();
+    assertThat(postBodyVerifier.hasDependency()).isTrue();
+    assertThat(postBodyVerifier.hasRequest()).isTrue();
+  }
+
+  class PostBodyVerifier {
+    boolean foundExceptionDoc = false;
+    boolean foundTraceDoc = false;
+    boolean foundDependency = false;
+    boolean foundRequest = false;
+
+    public void searchPostBody(String postBody) {
+      // Each post body is a list with a singular MonitoringDataPoint
+      List<MonitoringDataPoint> dataPoints = new ArrayList<>();
+      try {
+        JsonReader reader = JsonProviders.createReader(postBody);
+        dataPoints = reader.readArray(MonitoringDataPoint::fromJson);
+      } catch (IOException e) {
+        throw new IllegalStateException("Failed to parse post request body", e);
+      }
+
+      // Because the mock ping/posts should succeed, we should only have one MonitoringDataPoint per
+      // post
+      assertThat(dataPoints).hasSize(1);
+      MonitoringDataPoint dataPoint = dataPoints.get(0);
+
+      List<DocumentIngress> docs = dataPoint.getDocuments();
+      List<MetricPoint> metrics = dataPoint.getMetrics();
+
+      confirmDocsAreFiltered(docs);
+      confirmPerfCountersNonZero(metrics);
+      foundExceptionDoc = foundExceptionDoc || hasException(docs);
+      foundTraceDoc = foundTraceDoc || hasTrace(docs);
+      foundDependency = foundDependency || hasDependency(metrics);
+      foundRequest = foundRequest || hasRequest(metrics);
+    }
+
+    public boolean hasExceptionDoc() {
+      return foundExceptionDoc;
+    }
+
+    public boolean hasTraceDoc() {
+      return foundTraceDoc;
+    }
+
+    public boolean hasDependency() {
+      return foundDependency;
+    }
+
+    public boolean hasRequest() {
+      return foundRequest;
+    }
+
+    private void confirmDocsAreFiltered(List<DocumentIngress> docs) {
+      for (DocumentIngress doc : docs) {
+        assertThat(doc.getDocumentType()).isNotEqualTo(DocumentType.REMOTE_DEPENDENCY);
+        assertThat(doc.getDocumentType()).isNotEqualTo(DocumentType.REQUEST);
+      }
+    }
+
+    private boolean hasException(List<DocumentIngress> docs) {
+      for (DocumentIngress doc : docs) {
+        if (doc.getDocumentType().equals(DocumentType.EXCEPTION)
+            && ((Exception) doc).getExceptionMessage().equals("Fake Exception")) {
+          return true;
+        }
+      }
+      return false;
+    }
+
+    private boolean hasTrace(List<DocumentIngress> docs) {
+      for (DocumentIngress doc : docs) {
+        if (doc.getDocumentType().equals(DocumentType.TRACE)
+            && ((Trace) doc).getMessage().equals("This message should generate a trace")) {
+          return true;
+        }
+      }
+      return false;
+    }
+
+    private boolean hasDependency(List<MetricPoint> metrics) {
+      for (MetricPoint metric : metrics) {
+        String name = metric.getName();
+        double value = metric.getValue();
+        if (name.equals("\\ApplicationInsights\\Dependency Calls/Sec")) {
+          return value == 1;
+        }
+      }
+      return false;
+    }
+
+    private boolean hasRequest(List<MetricPoint> metrics) {
+      for (MetricPoint metric : metrics) {
+        String name = metric.getName();
+        double value = metric.getValue();
+        if (name.equals("\\ApplicationInsights\\Requests/Sec")) {
+          return value == 1;
+        }
+      }
+      return false;
+    }
+
+    private void confirmPerfCountersNonZero(List<MetricPoint> metrics) {
+      for (MetricPoint metric : metrics) {
+        String name = metric.getName();
+        double value = metric.getValue();
+        if (name.equals("\\Process\\Physical Bytes")
+            || name.equals("\\% Process\\Processor Time Normalized")) {
+          assertThat(value).isNotEqualTo(0);
+        }
+      }
+    }
   }
 
   @Environment(TOMCAT_8_JAVA_8)